requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohttp==3.9.1
aiolimiter==1.1.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
apscheduler==3.10.4
argcomplete==3.2.2
async-lru==2.0.4
async-timeout==4.0.3; python_version < "3.12.0"
asyncclick==8.1.7.1
asyncpg==0.29.0
attrs==23.2.0
bitarray==2.9.2
black==24.1a1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6; platform_system == "Windows" or sys_platform == "win32"
cytoolz==0.12.2; implementation_name == "cpython"
datamodel-code-generator==0.25.2
dnspython==2.5.0; python_version ~= "3.11"
email-validator==2.1.0.post1; python_version ~= "3.11"
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.0
eth-typing==4.0.0
eth-utils==3.0.0
frozenlist==1.4.1
genson==1.2.2
hexbytes==0.3.1
idna==3.6
inflect==5.6.2
iso8601==1.1.0
isort==5.13.2
jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.3.0
markupsafe==2.1.4
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
orjson==3.9.12
packaging==23.2
parsimonious==0.9.0
pathspec==0.12.1
platformdirs==4.1.0
prometheus-client==0.19.0
protobuf==4.25.2
pyarrow==14.0.2
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-core==2.14.6
pyhumps==3.8.0
pypika-tortoise==0.1.6
pysignalr==0.2.0
python-dotenv==1.0.1
pytz==2023.3.post1
pyunormalize==15.1.0
pywin32==306; platform_system == "Windows" or sys_platform == "win32"
pyyaml==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.17.1
ruamel-yaml==0.18.5
ruamel-yaml-clib==0.2.8; platform_python_implementation == "CPython" and python_version < "3.13"
sentry-sdk==1.39.2
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
strict-rfc3339==0.7
survey==4.5.4
tabulate==0.9.0
toolz==0.12.1; implementation_name == "pypy" or implementation_name == "cpython"
tortoise-orm==0.20.0
typing-extensions==4.9.0
tzdata==2023.4; platform_system == "Windows"
tzlocal==5.2
urllib3==2.1.0
web3==6.11.2
websockets==10.4
yarl==1.9.4
